+++ internal/deposit_service/deposit_relayer_service.go
@@ -116,7 +116,7 @@ func DepositRelayer(ctx context.Context, cfg *configs.Config, log logger.Logger,
 		panic(fmt.Sprintf("Unexpected transaction status: %d. Transaction Hash: %v", receipt.Status, receipt.TxHash.Hex()))
 	}
 
-	err = updateEventBlockNumber(db, log, mDBApp.DepositsRelayedEvent, int64(*depositIndices.LastDepositRelayedEventInfo.BlockNumber))
+	err = updateEventBlockNumber(db, log, mDBApp.DepositsRelayedEvent, *depositIndices.LastDepositRelayedEventInfo.BlockNumber)
 	if err != nil {
 		panic(fmt.Sprintf("Failed to update event block number: %v", err.Error()))
 	}
